requirements.txt
earthengine-api>=1.5.19
descartes>=1.1.0
folium>=0.14.0
geojson>=3.2.0
geopandas>=1.0.1
google-auth>=2.35.0
matplotlib>=3.7.0
numpy>=1.24.0
opencv-python>=4.8.0
pandas>=2.1.1
pyee>=11.1.0
scikit-learn>=1.3.0
shapely>=2.0.7
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
